finish and clean up
