+++ src/lib/route-data/utils/index.ts
@@ -1,0 +1,2 @@
+export * from './routes';
+export * from './route-path';
